Refactor into utils
