+++ backend/__init__.py
@@ -1,2 +1,2 @@
-from backend.dates import schedule_reminders
-from backend.handle_input import handle_input
+from backend.utils import *
+from backend.popup_classes import *
